requirements.txt
requests
torch
pydantic
Pillow
boto3>=1.34.101